switch to individual test files
