+++ integration-tests/jest/find-workflow.integration-test.js
@@ -1,0 +1,5 @@
+const { registerTest } = require('./jest-utils')
+const { testFindWorkflow } = require('../tests/find-workflow')
+
+
+registerTest(testFindWorkflow)
