+++ integration-tests/jest/import-dockstore-workflow.integration-test.js
@@ -1,0 +1,5 @@
+const { registerTest } = require('./jest-utils')
+const { testImportDockstoreWorkflow } = require('../tests/import-dockstore-workflow')
+
+
+registerTest(testImportDockstoreWorkflow)
