+++ integration-tests/jest/register-user.integration-test.js
@@ -1,0 +1,5 @@
+const { registerTest } = require('./jest-utils')
+const { testRegisterUser } = require('../tests/register-user')
+
+
+registerTest(testRegisterUser)
