+++ integration-tests/jest/run-workflow.integration-test.js
@@ -1,0 +1,5 @@
+const { registerTest } = require('./jest-utils')
+const { testRunWorkflow } = require('../tests/run-workflow')
+
+
+registerTest(testRunWorkflow)
