+++ integration-tests/tests/register-user.js
